complete product crud thing | changes in db

+++ store/views.py
@@ -3,31 +3,36 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
 from .models import Product
-from .serializers import ProductSerializer
+from .serializers import ProductSerializer, ProductCreateSerializer
 
 
 @api_view(['GET', 'POST'])
 def product_list(request):
     if request.method == 'GET':
         products = Product.objects.all()
-        serializer = ProductSerializer(products, many=True)
+        serializer = ProductSerializer(products, many=True, context={'request': request})
         return Response(serializer.data) 
     elif request.method == 'POST':
-        serializer = ProductSerializer(data=request.data)
+        serializer = ProductCreateSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response({'message': 'Product created successfully'}, status=status.HTTP_201_CREATED)
         
 
-@api_view(['GET', 'PUT', 'DELETE'])
+@api_view(['GET', 'PATCH', 'DELETE'])
 def product_detail(request, pk):
     product = get_object_or_404(Product, pk=pk)
     if request.method == 'GET':
-        serializer = ProductSerializer(product)
+        serializer = ProductSerializer(product, context={'request': request})
         return Response(serializer.data)
-    elif request.method == 'PUT':
-        return Response('working update route', status=status.HTTP_200_OK)
+    elif request.method == 'PATCH':
+        serializer = ProductCreateSerializer(product, data=request.data, partial=True)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response({'message': 'Product updated successfully'}, status=status.HTTP_200_OK)
     elif request.method == 'DELETE':
-        result = product.delete();
-        return Response(result, status=status.HTTP_204_NO_CONTENT)
+        if product.orderitem_set.count() > 0: 
+            return Response({'error': 'Product can not be deleted because it is associated with an order item.'},status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        product.delete();
+        return Response({'message': 'Product deleted successfully.'}, status=status.HTTP_204_NO_CONTENT)
